# api/openapi.json
{
  "components": {
    "schemas": {
      "AudioAppend": {
        "description": "One chunk of recording audio: base64 PCM16, 24 kHz, mono.",
        "properties": {
          "audio": {
            "minLength": 1,
            "title": "Audio",
            "type": "string"
          }
        },
        "required": [
          "audio"
        ],
        "title": "AudioAppend",
        "type": "object"
      },
      "AudioRead": {
        "description": "The saved recording's length, in seconds (0 when there is none yet).",
        "properties": {
          "duration_s": {
            "title": "Duration S",
            "type": "number"
          }
        },
        "required": [
          "duration_s"
        ],
        "title": "AudioRead",
        "type": "object"
      },
...
      "DailyNoteCreate": {
        "description": "The client's local date and the folder to hold today's note (Inbox\nwhen omitted). The date must be a real calendar date, never a bare\nstring the user could not have meant.",
        "properties": {
          "date": {
            "title": "Date",
            "type": "string"
          },
          "folder_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Folder Id"
          }
        },
        "required": [
          "date"
        ],
        "title": "DailyNoteCreate",
        "type": "object"
      },
      "DailyNoteRead": {
        "description": "Today's note, plus whether this call created it \u2014 the client can\ntell \"opened today's note\" from \"made today's note\" in one response.",
        "properties": {
          "created": {
            "title": "Created",
            "type": "boolean"
          },
          "entry": {
            "$ref": "#/components/schemas/EntryRead"
          }
        },
        "required": [
          "entry",
          "created"
        ],
        "title": "DailyNoteRead",
        "type": "object"
      },
...
      "EntryCreate": {
        "description": "A new blank note. When `folder_id` is omitted the note lands in Inbox.",
        "properties": {
          "folder_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Folder Id"
          }
        },
        "title": "EntryCreate",
        "type": "object"
      },
...
      "EntryRead": {
        "description": "One full entry: the human notes, the transcript, and the note's\nimages. Enhanced versions are read from /entries/{id}/enhanced-versions.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "folder_id": {
            "title": "Folder Id",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "images": {
            "items": {
              "$ref": "#/components/schemas/OutlineImage"
            },
            "title": "Images",
            "type": "array"
          },
          "note_md": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note Md"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "transcript": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transcript"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "folder_id",
          "title",
          "note_md",
          "transcript",
          "images",
          "created_at",
          "updated_at"
        ],
        "title": "EntryRead",
        "type": "object"
      },
      "EntrySummary": {
        "description": "One history row \u2014 just enough to list, plus a one-line preview, its\ntags (for folder chips and filtering), and its images (for the folder\nview's thumbnails) \u2014 no extra requests.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "folder_id": {
            "title": "Folder Id",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "images": {
            "items": {
              "$ref": "#/components/schemas/OutlineImage"
            },
            "title": "Images",
            "type": "array"
          },
          "preview": {
            "title": "Preview",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "folder_id",
          "title",
          "preview",
          "tags",
          "images",
          "created_at",
          "updated_at"
        ],
        "title": "EntrySummary",
        "type": "object"
      },
      "EntryUpdate": {
        "description": "The fields a note edit can change. `folder_id` moves the note.",
        "properties": {
          "folder_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Folder Id"
          },
          "note_md": {
            "anyOf": [
              {
                "maxLength": 1000000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note Md"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "EntryUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "meetwrite API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/entries": {
      "get": {
        "operationId": "read_entries_entries_get",
        "parameters": [
          {
            "in": "query",
            "name": "folder_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Folder Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EntrySummary"
                  },
                  "title": "Response Read Entries Entries Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Entries",
        "tags": [
          "entries"
        ]
      },
      "post": {
        "operationId": "post_entry_entries_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/EntryCreate"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Entry",
        "tags": [
          "entries"
        ]
      }
    },
    "/entries/daily": {
      "post": {
        "description": "Today's note: the same call always lands on the same note for one\ndate \u2014 created on the first call, returned ever after.",
        "operationId": "post_daily_entry_entries_daily_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DailyNoteCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyNoteRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Daily Entry",
        "tags": [
          "entries"
        ]
      }
    },
    "/entries/{entry_id}": {
      "delete": {
        "operationId": "remove_entry_entries__entry_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Entry",
        "tags": [
          "entries"
        ]
      },
      "get": {
        "operationId": "read_entry_entries__entry_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Entry",
        "tags": [
          "entries"
        ]
      },
      "patch": {
        "operationId": "patch_entry_entries__entry_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntryUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Entry",
        "tags": [
          "entries"
        ]
      }
    },
    "/entries/{entry_id}/audio": {
      "get": {
        "operationId": "read_entry_audio_entries__entry_id__audio_get",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Entry Audio",
        "tags": [
          "entries"
        ]
      },
      "post": {
        "operationId": "append_entry_audio_entries__entry_id__audio_post",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AudioAppend"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Append Entry Audio",
        "tags": [
          "entries"
        ]
      }
    },
    "/entries/{entry_id}/audio/file": {
      "get": {
        "operationId": "read_entry_audio_file_entries__entry_id__audio_file_get",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Entry Audio File",
        "tags": [
          "entries"
        ]
      }
    },
...
    "/entries/{entry_id}/images": {
      "post": {
        "description": "Store one picture on the note. Only declared image types are taken,\nand anything over MAX_IMAGE_BYTES is refused before it is read.",
        "operationId": "post_entry_image_entries__entry_id__images_post",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_post_entry_image_entries__entry_id__images_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryImageRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Entry Image",
        "tags": [
          "entries"
        ]
      }
    },
    "/entries/{entry_id}/images/{image_id}": {
      "delete": {
        "description": "Delete one picture. The note keeps any markdown reference; the\npreview simply renders nothing for a missing image id.",
        "operationId": "remove_entry_image_entries__entry_id__images__image_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "image_id",
            "required": true,
            "schema": {
              "title": "Image Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Entry Image",
        "tags": [
          "entries"
        ]
      },
      "get": {
        "description": "The picture itself \u2014 served straight from the row with its own\ncontent type, so previews and thumbnails need no extra encoding.",
        "operationId": "read_entry_image_file_entries__entry_id__images__image_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "image_id",
            "required": true,
            "schema": {
              "title": "Image Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Entry Image File",
        "tags": [
          "entries"
        ]
      },
      "patch": {
        "description": "Rename one picture (null clears the title back to untitled).",
        "operationId": "patch_entry_image_entries__entry_id__images__image_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "image_id",
            "required": true,
            "schema": {
              "title": "Image Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntryImageTitleUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryImageRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Entry Image",
        "tags": [
          "entries"
        ]
      }
    },
    "/entries/{entry_id}/outline": {
      "get": {
        "description": "The compact sidebar outline: transcript presence and chat titles.",
        "operationId": "read_entry_outline_entries__entry_id__outline_get",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryOutline"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Entry Outline",
        "tags": [
          "entries"
        ]
      }
    },
...
    "/entries/{entry_id}/tags": {
      "get": {
        "operationId": "read_entry_tags_entries__entry_id__tags_get",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Entry Tags",
        "tags": [
          "entries"
        ]
      },
      "put": {
        "description": "Replace the note's tags with exactly this list (normalized).",
        "operationId": "put_entry_tags_entries__entry_id__tags_put",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagReplace"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Entry Tags",
        "tags": [
          "entries"
        ]
      }
    },
...
  }
}